# openapi.yaml
openapi: 3.0.3
info:
  title: LucidLayer API
  version: 0.2.0
  description: |
    LucidLayer Offchain API.

    This OpenAPI spec is the source of truth for the actual backend routes.
    It is used to generate SDK clients via Speakeasy.

    Route groups:
    - `/v1/*` — Passports, Match, Run, Receipts, Epochs, Payouts, Compute
    - `/health/*` — System health and dependency checks
    - `/api/agents/*` — AI Agent MMR, Planner, and Orchestrator
servers:
  - url: https://api.lucid.foundation
    description: Production server

# Speakeasy MCP tool generation configuration
...
paths:
...
  /v1/epochs/current:
    get:
      tags: [Epochs]
      operationId: lucid_get_current_epoch
      summary: Get current epoch
      x-speakeasy-mcp:
        name: lucid_get_current_epoch
        description: Get the current epoch.
      parameters:
        - in: query
          name: project_id
          required: false
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [success, epoch]
                properties:
                  success: { type: boolean }
                  epoch: { $ref: '#/components/schemas/Epoch' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /v1/epochs:
    get:
      tags: [Epochs]
      operationId: lucid_list_epochs
      summary: List epochs
      x-speakeasy-mcp:
        name: lucid_list_epochs
        description: List epochs with optional filtering and pagination.
      parameters:
        - in: query
          name: project_id
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer, minimum: 1 }
        - in: query
          name: per_page
          schema: { type: integer, minimum: 1, maximum: 100 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [success, epochs, pagination]
                properties:
                  success: { type: boolean }
                  epochs:
                    type: array
                    items: { $ref: '#/components/schemas/Epoch' }
                  pagination: { $ref: '#/components/schemas/Pagination' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    post:
      tags: [Epochs]
      operationId: lucid_create_epoch
      summary: Create epoch
      x-speakeasy-mcp:
        name: lucid_create_epoch
        description: Create a new epoch.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                project_id: { type: string }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                required: [success, epoch]
                properties:
                  success: { type: boolean }
                  epoch: { $ref: '#/components/schemas/Epoch' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /v1/epochs/{epoch_id}:
    get:
      tags: [Epochs]
      operationId: lucid_get_epoch
      summary: Get epoch
      x-speakeasy-mcp:
        name: lucid_get_epoch
        description: Get an epoch by epoch_id.
      parameters:
        - in: path
          name: epoch_id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [success, epoch]
                properties:
                  success: { type: boolean }
                  epoch: { $ref: '#/components/schemas/Epoch' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /v1/epochs/{epoch_id}/retry:
    post:
      tags: [Epochs]
      operationId: lucid_retry_epoch
      summary: Retry failed epoch
      x-speakeasy-mcp:
        name: lucid_retry_epoch
        description: Retry a failed epoch.
      parameters:
        - in: path
          name: epoch_id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [success, epoch]
                properties:
                  success: { type: boolean }
                  epoch:
                    type: object
                    required: [epoch_id, status]
                    properties:
                      epoch_id: { type: string }
                      status: { type: string }
        '400': { $ref: '#/components/responses/BadRequest' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /v1/epochs/{epoch_id}/verify:
    get:
      tags: [Epochs]
      operationId: lucid_verify_epoch
      summary: Verify epoch anchor
      x-speakeasy-mcp:
        name: lucid_verify_epoch
        description: Verify an epoch anchor on-chain.
      parameters:
        - in: path
          name: epoch_id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [success, valid]
                properties:
                  success: { type: boolean }
                  valid: { type: boolean }
                  on_chain_root: { type: string, nullable: true }
                  expected_root: { type: string, nullable: true }
                  tx_signature: { type: string, nullable: true }
                  error: { type: string, nullable: true }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /v1/epochs/{epoch_id}/transaction:
    get:
      tags: [Epochs]
      operationId: lucid_get_epoch_transaction
      summary: Get epoch anchoring transaction
      x-speakeasy-mcp:
        name: lucid_get_epoch_transaction
        description: Get epoch anchoring transaction details.
      parameters:
        - in: path
          name: epoch_id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [success, tx_signature]
                properties:
                  success: { type: boolean }
                  tx_signature: { type: string }
                  slot: { type: integer }
                  block_time: { type: integer }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /v1/receipts/commit-root:
    post:
      tags: [Epochs]
      operationId: lucid_commit_epoch_root
      summary: Commit epoch root
      x-speakeasy-mcp:
        name: lucid_commit_epoch_root
        description: Commit an epoch root to chain.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                project_id: { type: string }
                epoch_id: { type: string }
                force: { type: boolean }
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                type: object
                required: [success, epoch_id, root, tx]
                properties:
                  success: { type: boolean }
                  epoch_id: { type: string }
                  root: { type: string }
                  tx: { type: string }
                  leaf_count: { type: integer }
        '400': { $ref: '#/components/responses/BadRequest' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /v1/receipts/commit-roots-batch:
    post:
      tags: [Epochs]
      operationId: lucid_commit_epoch_roots_batch
      summary: Commit multiple epoch roots
      x-speakeasy-mcp:
        name: lucid_commit_epoch_roots_batch
        description: Commit multiple epoch roots.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [epoch_ids]
              properties:
                epoch_ids:
                  type: array
                  items: { type: string }
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                type: object
                required: [success, total, successful_count, failed_count, results]
                properties:
                  success: { type: boolean }
                  total: { type: integer }
                  successful_count: { type: integer }
                  failed_count: { type: integer }
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        success: { type: boolean }
                        epoch_id: { type: string }
                        root: { type: string }
                        signature: { type: string }
                        error: { type: string }
        '400': { $ref: '#/components/responses/BadRequest' }
        '500': { $ref: '#/components/responses/InternalServerError' }
...
  /v1/epochs/stats:
    get:
      tags: [Epochs]
      operationId: lucid_get_epoch_stats
      summary: Epoch statistics
      x-speakeasy-mcp:
        name: lucid_get_epoch_stats
        description: Get epoch statistics.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EpochStatsResponse'
        '500': { $ref: '#/components/responses/InternalServerError' }

  # ===========================================================================
  # Payouts
  # ===========================================================================
...
components:
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    NotFound:
      description: Not Found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    ServiceUnavailable:
      description: Service Unavailable
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
...
    Pagination:
      type: object
      required: [total, page, per_page, total_pages]
      properties:
        total: { type: integer }
        page: { type: integer }
        per_page: { type: integer }
        total_pages: { type: integer }

    # =========================================================================
    # Policy
    # =========================================================================
...
    ErrorResponse:
      type: object
      properties:
        success: { type: boolean }
        error: { type: string }
        message: { type: string }
        error_code: { type: string }
        details: {}

    # =========================================================================
    # Compute
    # =========================================================================
...
    Epoch:
      type: object
      required: [epoch_id, mmr_root, leaf_count, created_at, status]
      properties:
        epoch_id: { type: string }
        project_id: { type: string }
        mmr_root: { type: string }
        leaf_count: { type: integer }
        created_at: { type: integer }
        finalized_at: { type: integer }
        status:
          type: string
          enum: [open, anchoring, anchored, failed]
        chain_tx: { type: string }
        error: { type: string }
        start_leaf_index: { type: integer }
        end_leaf_index: { type: integer }
    EpochStatsResponse:
      type: object
      required: [success, stats]
      properties:
        success: { type: boolean }
        stats:
          type: object
          properties:
            total_epochs: { type: integer }
            total_receipts: { type: integer }
            by_status:
              type: object
              properties:
                open: { type: integer }
                anchoring: { type: integer }
                anchored: { type: integer }
                failed: { type: integer }
            avg_receipts_per_epoch: { type: number }
            last_anchor_time: { type: integer }
...